SubGhz: refactoring
