+++ applications/main/subghz/helpers/subghz_txrx.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "subghz_types.h"
+#include "subghz_txrx_callbacks.h"
 #include <lib/subghz/subghz_worker.h>
 #include <lib/subghz/subghz_setting.h>
 #include <lib/subghz/receiver.h>
@@ -80,88 +81,3 @@ void subghz_txrx_set_debug_pin_state(SubGhzTxRx* instance, bool state);
 bool subghz_txrx_get_debug_pin_state(SubGhzTxRx* instance);
 
 SubGhzReceiver* subghz_txrx_get_receiver(SubGhzTxRx* instance); // TODO use only in DecodeRaw
-
-//#############Create  new Key##############
-bool subghz_txrx_gen_data_protocol(
-    void* context,
-    const char* preset_name,
-    uint32_t frequency,
-    const char* protocol_name,
-    uint64_t key,
-    uint32_t bit);
-
-bool subghz_txrx_gen_data_protocol_and_te(
-    SubGhzTxRx* instance,
-    const char* preset_name,
-    uint32_t frequency,
-    const char* protocol_name,
-    uint64_t key,
-    uint32_t bit,
-    uint32_t te);
-
-bool subghz_scene_set_type_submenu_gen_data_keeloq(
-    SubGhzTxRx* txrx,
-    const char* preset_name,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint16_t cnt,
-    const char* manufacture_name);
-
-bool subghz_scene_set_type_submenu_gen_data_keeloq_bft( //TODO rename
-    void* context,
-    const char* preset_name,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint16_t cnt,
-    uint32_t seed,
-    const char* manufacture_name);
-
-bool subghz_scene_set_type_submenu_gen_data_nice_flor( //TODO rename
-    void* context,
-    const char* preset_name,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint16_t cnt,
-    bool nice_one);
-
-bool subghz_scene_set_type_submenu_gen_data_faac_slh( //TODO rename
-    void* context,
-    const char* preset_name,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint16_t cnt,
-    uint32_t seed,
-    const char* manufacture_name);
-
-bool subghz_scene_set_type_submenu_gen_data_alutech_at_4n( //TODO rename
-    void* context,
-    const char* preset_name,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint16_t cnt);
-
-bool subghz_scene_set_type_submenu_gen_data_somfy_telis( //TODO rename
-    void* context,
-    const char* preset_name,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint16_t cnt);
-
-bool subghz_txrx_gen_secplus_v2_protocol(
-    SubGhzTxRx* txrx,
-    const char* name_preset,
-    uint32_t frequency,
-    uint32_t serial,
-    uint8_t btn,
-    uint32_t cnt);
-
-bool subghz_txrx_gen_secplus_v1_protocol(
-    SubGhzTxRx* txrx,
-    const char* name_preset,
-    uint32_t frequency);
