+++ applications/main/subghz/helpers/subghz_txrx_callbacks.h
@@ -1,0 +1,2 @@
+#pragma once
+typedef void (*SubGhzTxRxNeedSaveCallback)(void* context);
