+++ applications/main/subghz/scenes/subghz_scene_receiver.c
@@ -90,27 +90,26 @@ static void subghz_scene_add_to_history_callback(
     SubGhzProtocolDecoderBase* decoder_base,
     void* context) {
     furi_assert(context);
-
     SubGhz* subghz = context;
-
+    SubGhzHistory* history = subghz->history;
     FuriString* item_name = furi_string_alloc();
     FuriString* item_time = furi_string_alloc();
-    uint16_t idx = subghz_history_get_item(subghz->history);
+    uint16_t idx = subghz_history_get_item(history);
 
     SubGhzRadioPreset preset = subghz_txrx_get_preset(subghz->txrx);
-    if(subghz_history_add_to_history(subghz->history, decoder_base, &preset)) {
+    if(subghz_history_add_to_history(history, decoder_base, &preset)) {
         furi_string_reset(item_name);
         furi_string_reset(item_time);
 
         subghz->state_notifications = SubGhzNotificationStateRxDone;
 
-        subghz_history_get_text_item_menu(subghz->history, item_name, idx);
-        subghz_history_get_time_item_menu(subghz->history, item_time, idx);
+        subghz_history_get_text_item_menu(history, item_name, idx);
+        subghz_history_get_time_item_menu(history, item_time, idx);
         subghz_view_receiver_add_item_to_menu(
             subghz->subghz_receiver,
             furi_string_get_cstr(item_name),
             furi_string_get_cstr(item_time),
-            subghz_history_get_type_protocol(subghz->history, idx));
+            subghz_history_get_type_protocol(history, idx));
 
         subghz_scene_receiver_update_statusbar(subghz);
     }
@@ -122,13 +121,14 @@ static void subghz_scene_add_to_history_callback(
 
 void subghz_scene_receiver_on_enter(void* context) {
     SubGhz* subghz = context;
+    SubGhzHistory* history = subghz->history;
 
     FuriString* item_name = furi_string_alloc();
     FuriString* item_time = furi_string_alloc();
 
     if(subghz_rx_key_state_get(subghz) == SubGhzRxKeyStateIDLE) {
         subghz_txrx_set_preset(subghz->txrx, "AM650", subghz->last_settings->frequency, NULL, 0);
-        subghz_history_reset(subghz->history);
+        subghz_history_reset(history);
         subghz_rx_key_state_set(subghz, SubGhzRxKeyStateStart);
     }
 
@@ -137,16 +137,16 @@ void subghz_scene_receiver_on_enter(void* context) {
 
     //Load history to receiver
     subghz_view_receiver_exit(subghz->subghz_receiver);
-    for(uint8_t i = 0; i < subghz_history_get_item(subghz->history); i++) {
+    for(uint8_t i = 0; i < subghz_history_get_item(history); i++) {
         furi_string_reset(item_name);
         furi_string_reset(item_time);
-        subghz_history_get_text_item_menu(subghz->history, item_name, i);
-        subghz_history_get_time_item_menu(subghz->history, item_time, i);
+        subghz_history_get_text_item_menu(history, item_name, i);
+        subghz_history_get_time_item_menu(history, item_time, i);
         subghz_view_receiver_add_item_to_menu(
             subghz->subghz_receiver,
             furi_string_get_cstr(item_name),
             furi_string_get_cstr(item_time),
-            subghz_history_get_type_protocol(subghz->history, i));
+            subghz_history_get_type_protocol(history, i));
         subghz_rx_key_state_set(subghz, SubGhzRxKeyStateAddKey);
     }
     furi_string_free(item_name);
